Refactor: move functions to clowder_utilities
