+++ clowder/clowder_yaml.py
@@ -1,9 +1,13 @@
 """clowder.yaml parsing and functionality"""
-import os, subprocess, sys, yaml
+import os, yaml
 from termcolor import colored
 from clowder.group import Group
 from clowder.source import Source
-from clowder.utility.clowder_utilities import print_exiting
+from clowder.utility.clowder_utilities import (
+    _forall_run,
+    _validate_yaml,
+    print_exiting
+)
 
 class ClowderYAML(object):
     """Class encapsulating project information from clowder.yaml"""
@@ -152,45 +156,3 @@ class ClowderYAML(object):
                     valid = False
         if not valid:
             print_exiting()
-
-# Disable errors shown by pylint for no specified exception types
-# pylint: disable=W0702
-# Disable errors shown by pylint for statements which appear to have no effect
-# pylint: disable=W0104
-def _validate_yaml(parsed_yaml):
-    """Load clowder from yaml file"""
-    try:
-        parsed_yaml['defaults']['ref']
-        parsed_yaml['defaults']['remote']
-        parsed_yaml['defaults']['source']
-
-        for source in parsed_yaml['sources']:
-            source['name']
-            source['url']
-
-        for group in parsed_yaml['groups']:
-            group['name']
-            for project in group['projects']:
-                project['name']
-                project['path']
-    except:
-        print('')
-        clowder_output = colored('clowder.yaml', 'cyan')
-        print(clowder_output + ' appears to be invalid')
-        print_exiting()
-
-def _forall_run(command, directories):
-    """Run command in all directories"""
-    sorted_paths = sorted(set(directories))
-    paths = [p for p in sorted_paths if os.path.isdir(p)]
-    for path in paths:
-        running_output = colored('Running command', attrs=['underline'])
-        command_output = colored(command, attrs=['bold'])
-        print(running_output + ': ' + command_output)
-        directory_output = colored('Directory', attrs=['underline'])
-        path_output = colored(path, 'cyan')
-        print(directory_output + ': ' + path_output)
-        subprocess.call(command.split(),
-                        cwd=path)
-        print('')
-    sys.exit()
