+++ clowder/utility/clowder_utilities.py
@@ -1,5 +1,5 @@
 """Clowder utilities"""
-import emoji, os, sys
+import emoji, os, subprocess, sys
 from git import Repo
 from termcolor import colored, cprint
 from clowder.utility.git_utilities import (
@@ -109,3 +109,46 @@ def validate_repo_state(repo_path):
     if not os.path.isdir(os.path.join(repo_path, '.git')):
         return True
     return not git_is_dirty(repo_path)
+
+def _forall_run(command, directories):
+    """Run command in all directories"""
+    sorted_paths = sorted(set(directories))
+    paths = [p for p in sorted_paths if os.path.isdir(p)]
+    for path in paths:
+        running_output = colored('Running command', attrs=['underline'])
+        command_output = colored(command, attrs=['bold'])
+        print(running_output + ': ' + command_output)
+        directory_output = colored('Directory', attrs=['underline'])
+        path_output = colored(path, 'cyan')
+        print(directory_output + ': ' + path_output)
+        subprocess.call(command.split(),
+                        cwd=path)
+        print('')
+    # Exit early to prevent printing extra newline
+    sys.exit()
+
+# Disable errors shown by pylint for no specified exception types
+# pylint: disable=W0702
+# Disable errors shown by pylint for statements which appear to have no effect
+# pylint: disable=W0104
+def _validate_yaml(parsed_yaml):
+    """Load clowder from yaml file"""
+    try:
+        parsed_yaml['defaults']['ref']
+        parsed_yaml['defaults']['remote']
+        parsed_yaml['defaults']['source']
+
+        for source in parsed_yaml['sources']:
+            source['name']
+            source['url']
+
+        for group in parsed_yaml['groups']:
+            group['name']
+            for project in group['projects']:
+                project['name']
+                project['path']
+    except:
+        print('')
+        clowder_output = colored('clowder.yaml', 'cyan')
+        print(clowder_output + ' appears to be invalid')
+        print_exiting()
